Day - 6: Further Optimisation and Log Fixes
Okay so I managed to finally fix the function of clearing the value of the table once the messages pop off in the channel and it will clear the value once its done. Hopefully I can instead clean the code instead of making them a jumble mumble of this mess.